debugged diffusion error in spotlight!!

+++ include/Intersection.h
@@ -5,8 +5,8 @@ class Shape;
 
 class Intersection{
     public:
-        Intersection(Shape *shape, Ray hit, int t) : shape(shape), hit(hit), t(t){};
+        Intersection(Shape *shape, Ray hit, float t) : shape(shape), hit(hit), t(t){};
         const Shape* const shape;
         const Ray hit;
-        const int t; //the t that solves the intersection equasion
+        const float t; //the t that solves the intersection equasion
 };
